Updates dbscan.cpp
#30 #31 Modifies `run_dbscan_algo_uf_mpi_interleaved()` function. Sets
up a “short circuit” in the “main loop” if the SNG Alg is used. If the
“seed point” has been seen before, the loop moves to the next
iteration. After checking if the `mints` requirement has been met, if
the SNG Algorithm is used is checked. If True, a while loop is used
till the `triage` vector is empty.
Modifies `unionize_neighborhood()` function. For the remote points, if
the SNG Alg is used, if the point has been seen before, the loop moves
to the next iteration.