Updates dbscan.cpp
#40 Modifies the ClusteringAlgo::modify_status_vectors() function. The
`assessed` and `triage` vectors are modified more easily. A new
variable, kdtree2_result_vector newNe, is now used to replace the local
neighborhood `ne` vector, to make use of the “neighborhood pruning”
technique. This isn’t complete though. This is just a benchmark. The
remote points now need to be worked on.